+++ app/Models/student/Prospect.php
@@ -1,12 +1,14 @@
 <?php
 
-namespace App\Models\follow;
+namespace App\Models\student;
 
 use Illuminate\Database\Eloquent\Model;
 
 class Prospect extends Model
 {
     protected $fillable = [
-        'company_name', 'phone_contact', 'email_contact',
+        'company_name',
+        'phone_contact',
+        'email_contact',
     ];
 }
